Integrate front-end and API
Simplified running the app in development and production mode. Express serves the /build folder and `npm start` simply builds the
front-end and runs the API.
