+++ scripts/run-build.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const build = require('./build');
+
+build();
