+++ src/server/app.js
@@ -1,10 +1,13 @@
+'use strict';
+
 const express = require('express');
 const path = require('path');
-const favicon = require('serve-favicon');
 const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 
+const BUILD_PATH = path.join(__dirname, '..', '..', 'build');
+const INDEX_PATH = path.join(BUILD_PATH, 'index.html');
 const api = require('./controllers/api');
 
 const app = express();
@@ -23,7 +26,13 @@ app.use(function(req, res, next) {
   next();
 });
 
+
+
+app.use(express.static(path.join(BUILD_PATH)));
 app.use('/api', api);
+app.get('*', function(req, res){
+  res.sendFile(INDEX_PATH);
+})
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -37,7 +46,7 @@ app.use(function(req, res, next) {
 // development error handler
 // will print stacktrace
 if (app.get('env') === 'development') {
-  app.use(function(err, req, res, next) {
+  app.use(function(err, req, res) {
     res.status(err.status || 500);
     res.json({
       message: err.message,
@@ -48,7 +57,7 @@ if (app.get('env') === 'development') {
 
 // production error handler
 // no stacktraces leaked to user
-app.use(function(err, req, res, next) {
+app.use(function(err, req, res) {
   res.status(err.status || 500);
   res.json({
     message: err.message,
